correction 01

+++ libft/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: aldiaz-c <aldiaz-c@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/05 19:02:59 by aldiaz-c          #+#    #+#             */
-/*   Updated: 2022/05/05 17:23:32 by aldiaz-c         ###   ########.fr       */
+/*   Updated: 2022/05/16 17:32:08 by aldiaz-c         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i])
